+++ libgringo/src/input/program.cc
@@ -98,7 +98,7 @@ void Program::rewrite(Defines &defs, Logger &log) {
             args.emplace_back(gen.uniqueVar(param.first, 0, "#Inc"));
             incDefs.add(param.first, param.second, get_clone(args.back()), false, log);
         }
-        sigs_.push(Sig(block.name, args.size(), false));
+        sigs_.push(Sig(block.name, static_cast<uint32_t>(args.size()), false));
         UTerm blockTerm(args.empty()
             ? (UTerm)make_locatable<ValTerm>(block.loc, Symbol::createId(block.name))
             : make_locatable<FunctionTerm>(block.loc, block.name, get_clone(args)));
